plotcell is mostly cleaned up.
we had to make sure that we specified the tick locations before assigning tickc labels! so now we just tally properties and apply them at the end. This, conveniently, is also a step towards not assigning axes until the call to draw(), and towards using @property instead of a style() call.